refactor(probe): migrate rip and ripng to per-protocol plugin
